[feat] push-setting / 푸시알림 설정 서버 붙이기 (#65)

+++ feature/push-setting/src/main/java/com/photosurfer/android/push_setting/fragments/SelectTagFragment.kt
@@ -28,7 +28,7 @@ class SelectTagFragment : BaseFragment<FragmentSelectTagBinding>(R.layout.fragme
         selectTagListAdapter.submitList(pushSettingViewModel.wholeTagList.value)
     }
 
-    private fun updateTempRepresentTagList(tempList: MutableList<Long>) {
+    private fun updateTempRepresentTagList(tempList: MutableList<Int>) {
         pushSettingViewModel.updateTempRepresentTagList(tempList)
     }
 }
